+++ Gusboy/Core/RAM.cs
@@ -702,7 +702,7 @@
         }
 
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        public ushort GetShort(int address) => (ushort)((this[address + 1] << 8) + this[address]);
+        public int GetShort(int address) => (this[address + 1] << 8) + this[address];
 
         internal void ClearRAM()
         {
